requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.6.0
Markdown==3.2.2
mysqlclient==1.4.6
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
zipp==3.1.0
